+++ src/screens/Home and Profile/HomeScreen.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { View, Text, Button, ActivityIndicator, StyleSheet } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { Colors } from '../utils/theme';
+import { Colors } from '../../utils/theme';
 
 const HomeScreen = ({ navigation }) => {
